+++ pac4j-core/src/main/java/org/pac4j/core/util/Announce.java
@@ -19,9 +19,12 @@ public final class Announce {
             LOGGER.warn("\u26A0 + v6.5.0:");
             LOGGER.warn("\u26A0  - the `pac4j-gae`, `pac4j-couch` and `pac4j-springboot` modules will be removed");
             LOGGER.warn("\u26A0  - the OSGi and shading Maven phases will be removed");
+            LOGGER.warn("\u26A0  - the JSON serializer is the default one for `(Ldap|Db|Mongo)ProfileService`");
             LOGGER.warn("\u26A0 + v7.0.0:");
             LOGGER.warn("\u26A0  - the `pac4j-config` module will be removed");
             LOGGER.warn("\u26A0  - the `pac4j-javaee` module will likely be removed");
+            LOGGER.warn("\u26A0  - all deprecated elements will be removed");
+            LOGGER.warn("\u26A0  - the 'legcay mode' of the `(Ldap|Db|Mongo)ProfileService` will be removed");
             displayed = true;
         }
     }